Ignore the BOM if found at the start of a source
The GLSL spec specifies that the source encoding must be UTF-8, which
makes the BOM rather pointless. However, it could not possibly be
mistaken for valid code, so there is no possibility of breaking a
valid program by ignoring it.

> Data streams (or files) that begin with the U+FEFF byte order mark
> are likely to contain Unicode characters. It is recommended that
> applications sending or receiving untyped data streams of coded
> characters use this signature. If other signaling methods are used,
> signatures should not be employed.

Our signalling method is the GLSL standard, which specifies there is
only one possible format: UTF-8. Arguably, the BOM should therefore be
interpreted as a ZWNBS, but that's not a valid character in GLSL, so
we can be friendly and optionally allow the BOM.