Switch non-public guards to #pragma once
There are already a few headers that use #pragma once, so clearly there
aren't any compatability problems introduced with this move. The current
guard symbols are inconsistent and frequently use reserved tokens.

I left the headers in Include and Public as-is, just in case someone
depends on them. That would be weird, but hey. If we want to fix them
too, `guard2once -rs .` will do it.

If for some reason `#pragma once` turns out to be a problem, all the
headers can be switched back with `once2guard -rp 'path -1 | upper' .`
As a bonus, they'll all be given a consistent naming scheme in the
process.